// frontend/src/external/client/api/orval/generated/models/modelsCreateArticleRequest.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * inselfy API
 * OpenAPI spec version: 1.0.0
 */

/**
 * 記事作成リクエスト
 */
export interface ModelsCreateArticleRequest {
  /**
     * タイトル
     * @maxLength 200
     */
  title: string;
  /**
     * 本文
     * @maxLength 100000
     */
  body: string;
  /** 有料記事か */
  isPaid: boolean;
  /**
     * 価格（円）
     * @minimum 0
     * @maximum 1000000
     */
  priceYen: number;
  /**
     * カバー画像URL
     * @nullable
     */
  coverImageUrl?: string | null;
  /**
     * タグ（各タグは50文字以内）
     * @maxItems 10
     */
  tags: string[];
}